# Small vector optimization