Add SVO to matrix classes directly, instead of using the vendor-dependant std::basic_string.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/mrpt/mrpt/934)
<!-- Reviewable:end -->