@jolting 2019-07-01
@jlblancoc I found out my laptop was thermal throttling during tests. Not the best thing to benchmark with. However, I think one of the map resize benchmarks might have been improved by this.