@jlblancoc 2019-07-03
Well, test added in fb4b77ab6fd9599bda265853280c42ba70744618

The gain is still marginal (perhaps the compiler is optimizing out something since the matrices are not actually being used?), but it's something, I would vote to keep the svo thing: 

current master: 
```
mrpt-performance -c "matrix: vector"

matrix: vector of, resize(10) dyn[double] 4x4               1.99 us
matrix: vector of, resize(100) dyn[double] 4x4              7.98 us
matrix: vector of, resize(1000) dyn[double] 4x4             81.06 us
matrix: vector of, resize(10) dyn[double] 5x5               809.80 ns
matrix: vector of, resize(100) dyn[double] 5x5              8.46 us
matrix: vector of, resize(1000) dyn[double] 5x5             80.69 us
matrix: vector of, resize(10) fix[double] 4x4               140.50 ns
matrix: vector of, resize(100) fix[double] 4x4              301.30 ns
matrix: vector of, resize(1000) fix[double] 4x4             2.56 us
matrix: vector of, resize(10) fix[double] 5x5               201.30 ns
matrix: vector of, resize(100) fix[double] 5x5              906.70 ns
matrix: vector of, resize(1000) fix[double] 5x5             8.44 us
```

this PR:
```
matrix: vector of, resize(10) dyn[double] 4x4               2.04 us
matrix: vector of, resize(100) dyn[double] 4x4              8.22 us
matrix: vector of, resize(1000) dyn[double] 4x4             75.40 us
matrix: vector of, resize(10) dyn[double] 5x5               863.00 ns
matrix: vector of, resize(100) dyn[double] 5x5              7.36 us
matrix: vector of, resize(1000) dyn[double] 5x5             79.65 us
matrix: vector of, resize(10) fix[double] 4x4               146.90 ns
matrix: vector of, resize(100) fix[double] 4x4              298.10 ns
matrix: vector of, resize(1000) fix[double] 4x4             2.38 us
matrix: vector of, resize(10) fix[double] 5x5               207.00 ns
matrix: vector of, resize(100) fix[double] 5x5              898.90 ns
matrix: vector of, resize(1000) fix[double] 5x5             8.72 us
```

(Timing for "fixed" matrices shouldn't change between the tests, changes are just "random")