@jlblancoc 2019-07-11
Merged, thanks @jolting for the great alternative to `std::basic_string`, with well-defined, predictable behavior. Cool!